+++ CHANGES.rst
@@ -19,6 +19,7 @@ Unreleased
 * Add methods :meth:`.add_calendar` and :meth:`.add_community_list`.
 * Add methods :meth:`.add_image_widget` and :meth:`.upload_image`.
 * Add method :meth:`.add_custom_widget`.
+* Add method :meth:`.add_post_flair_widget`.
 
 **Changed**